+++ src/main/FullStandardUser.java
@@ -6,7 +6,7 @@ package main;
 public class FullStandardUser extends AbstractUser {
 
 
-    public FullStandardUser(String username, Float credit) {
+    public FullStandardUser(String username, double credit) {
         super(username);
         this.accountBalance = credit;
         this.type = "FS";
@@ -22,7 +22,7 @@ public class FullStandardUser extends AbstractUser {
     //THIS DOES NOT FOLLOW THE RIGHT FORMAT
 
     @Override
-    public void create(String username, String type, Float credit, Application application){
+    public void create(String username, String type, double credit, Application application){
         System.out.println(this.username + " cannot create another user.");
     }
 
@@ -31,7 +31,7 @@ public class FullStandardUser extends AbstractUser {
     //THIS DOES NOT FOLLOW THE RIGHT FORMAT
 
     @Override
-    public void delete(AbstractUser user, float amount){
+    public void delete(AbstractUser user, double amount){
         System.out.println(this.username + " cannot delete another user.");
     }
 
@@ -46,7 +46,7 @@ public class FullStandardUser extends AbstractUser {
     //THIS DOES NOT FOLLOW THE RIGHT FORMAT
 
     @Override
-    public boolean refund(AbstractUser buyer, AbstractUser seller, float amount){
+    public boolean refund(AbstractUser buyer, AbstractUser seller, double amount){
         System.out.println(this.username + " cannot issue a refund.");
         return false;
     }
